prometheus: Refactor the components and add docs
this patch refactors the components from common.tsx
into separate components. Also adds documentation
wherever necessary.
