+++ prometheus/src/setupTests.js
@@ -1,0 +1,5 @@
+import '@testing-library/jest-dom';
+import { TextDecoder, TextEncoder } from 'util';
+
+global.TextEncoder = TextEncoder;
+global.TextDecoder = TextDecoder;
